+++ audit_functions.php
@@ -1,8 +1,7 @@
 <?php
 
-function audit_user_can_purge() {
-	return api_plugin_user_realm_auth('audit_manage.php') ||
-		api_plugin_user_realm_auth('audit_purge.php');
+function audit_user_is_admin() {
+	return api_plugin_user_realm_auth('audit_manage.php');
 }
 
 function audit_process_page_data($page, $drop_action, $selected_items) {